% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcObjFnPseudo.R
\name{calcObjPCApseudo}
\alias{calcObjPCApseudo}
\title{calcObjPCApseudo() Function}
\usage{
calcObjPCApseudo(
  resDir,
  newAnnotation,
  newAnnotationRscriptName,
  expCondSepName,
  expCondName2change,
  subsetCond,
  subsetClusters
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{expCondSepName}{character string, user defined name either to be 'org' or any character string}

\item{expCondName2change}{if above 'expCondSepName' is defined not as 'org', provide the name to be changed}
}
\value{
a SingleCellExperiment where 3 methods functional pseudotime analysis results are saved in colData
}
\description{
calcObjPCApseudo() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
calcObjPCApseudo()
}
\keyword{calcObjPCApseudo}
