% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compAnno.R
\name{compAnno}
\alias{compAnno}
\title{compAnno() Function}
\usage{
compAnno(seuratObj, seuratObj2Comp)
}
\arguments{
\item{obj}{either a Seurat or SingleCellExperiment object.}

\item{slingshotclusterLabels}{cluster used for slingshot with 3 options: NULL, GMM, Seurat_clusters, by default NULL.}

\item{resSave}{whether to save the analysis results into Rdata, by default 'FALSE'}

\item{resFnamePrefix}{if 'resSave = T', the Rdata file name prefix, by default 'TEST'}
}
\value{
results in cell number and relative percentage of idents cells.
}
\description{
compAnno() Function
}
\details{
This function is used to perform functional pseudo time analysis via PCA, Diffusion Map, and slingshot
}
\examples{
compAnno()
}
\keyword{compAnno}
