% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReadin.R
\name{countReadin}
\alias{countReadin}
\title{countReadin() Function}
\usage{
countReadin(
  cellrangerResList,
  resDirName,
  genomeSpecies,
  minCells,
  minFeatures,
  doubletsRmOn,
  doubletsRmMethods,
  doubletsResDir,
  mtFiltering,
  mtPerCutoff
)
}
\arguments{
\item{cellrangerResList}{list item, to specify full path to cellranger analysis results for each sample.}

\item{resDirName}{optional, define folder/directory name where integration analysis results will be saved}

\item{genomeSpecies}{To be added}

\item{minCells}{optional, default = 3, used in creating seurat object to indicate the minumn detected number of cells to be included.}

\item{minFeatures}{optional, default = 200, used in creating seurat object to indicate the minimum detected number of gene features (genes) of included cells.}

\item{doubletsRmOn}{required, default = T}

\item{doubletsRmMethods}{To be added}

\item{doubletsResDir}{To be added}

\item{mtFiltering}{default = F, logical, whehter to filter mitochondrial content}

\item{mtPerCutoff}{optional, if mtFiltering = T, default = 20, indicating the percentage cut-off of mitochondrial content}
}
\value{
results of analysis
}
\description{
countReadin() Function
}
\details{
This function is used to readin cellranger results into seurat object by defining w/wo doublets removal
}
\examples{
findDoublets()
}
\keyword{countReadin}
