% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDoublets.R
\name{findDoublets}
\alias{findDoublets}
\title{findDoublets() Function}
\usage{
findDoublets(
  cellrangerResList,
  genomeSpecies,
  doubletDeconRhop,
  doubletDeconPMF,
  doubletDeconNoCore,
  filenamePrefix
)
}
\arguments{
\item{cellrangerResList}{list including full path to cellranger analysis results for different samples.}

\item{genomeSpecies}{genome species, supporting human, mouse, and rate so far.}
}
\value{

}
\description{
This function allows you to find and estimate doublets with DoubletDecon for the provide iput 'cellrangerResList'
}
\examples{
findDoublets()
}
\keyword{findDoublets}
