% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellChatRes.R
\name{getCellChatRes}
\alias{getCellChatRes}
\title{getCellChatRes() Function}
\usage{
getCellChatRes(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  genomeSpecies = "human",
  selected.DB = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  expCond = NULL,
  min.cells.filtering = 10,
  debug = F
)
}
\arguments{
\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{rds}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{genomeSpecies}{specify sample's genome species, by default 'human', currently supporting human, mouse, and rate.}

\item{selected.DB}{whether to conduct cellChat analysis on the selected pathway, a total of 3 options to choose, they are 'Cell-Cell Contact', 'ECM-Receptor' and 'Secreted Signaling'.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{expCondCheckFname}{prefix of saved rds file name.}

\item{expCond}{specify the specific experimental condition for cellchat analysis, if not specified, all experimental conditions will be performed .}

\item{min.cells.filtering}{the minimum number of cells in the group used for analysis.}
}
\value{
cellChat analysis result
}
\description{
getCellChatRes() Function
}
\details{
This function is used to conduct cellChat analysis on the selected expCond levels.
}
\examples{
getCellChatRes(rds, expCondCheck='sample/idents/expCond*')
}
\keyword{getCellChatRes}
