% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChiResults.R
\name{getChiResults}
\alias{getChiResults}
\title{getChiResults() Function}
\usage{
getChiResults(
  perTable,
  expCondCheck = NULL,
  chisqMonteCarlo = F,
  plotLegendPer = 0.4,
  plotCex = 2,
  legendCex = 2,
  legendDistance = 0.3,
  axisLableDistance = 1,
  topTextDegree = 90,
  plotFnamePrefix = "test",
  plotVertical = F,
  plotWidth = 7,
  plotHeight = 7,
  resReturn = F,
  debug = F
)
}
\arguments{
\item{perTable}{cell number decomposion table, where row represent *** and column represent.}

\item{expCondCheck}{specify which columns in a regular expression pattern to be test and plot.}

\item{chisqMonteCarlo}{a logical indicating whether to compute p-values by Monte Carlo simulation, by default = Fasle.}

\item{plotLegendPer}{plot specification on the legend bar, specify the percentage of legend bar to occupy the entire plot, by default = 0.4.}

\item{plotCex}{plot specification on the cex size of the plot, by default = 2.}

\item{legendCex}{plot specification on the cex size of the legend, by default = 2.}

\item{legendDistance}{plot specification on the distance between legend text and legend bar, by default = 0.3.}

\item{axisLableDistance}{plot specification on the distance between axis label text and plot, by default = 1.}

\item{topTextDegree}{plot specification on the rotation degree of top label text, by default = 90 (vertical align).}

\item{plotFnamePrefix}{plot specification on the plot file name prefix, by default a plot named 'test.pdf' will be generated.}

\item{plotVertical}{plot specification on the plot display whether to rotate column and row displays on the plot.}

\item{plotWidth}{plot specification on the plot width size, by default = 7.}

\item{plotHeight}{plot specification on the plot height size, by default = 7.}

\item{resReturn}{specify whether chi-square test results to be returned in a list format, by default = False.}

\item{debug}{by default False, for development purpose to check the analysis progress.}
}
\value{
if specified to return, a list of 3 items can be returned, they are 'chiRes', 'residual' and 'contrib'.
}
\description{
getChiResults() Function
}
\details{
This function is used to perform chi-square test on the cell number decomposition table to generate the corresponding contribution plot.
}
\examples{
getChiResults(perTable)
}
\keyword{getChiResults}
