% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterExpCondDE.R
\name{getClusterExpCondDe}
\alias{getClusterExpCondDe}
\title{getClusterExpCondDe() Function}
\usage{
getClusterExpCondDe(
  resDir = NULL,
  rdsFname = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  compGroup,
  deMethod = "wilcox",
  pAdjValCutoff = 0.05,
  topNo = 10
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{compGroup}{group names used for comparisions, e.g. 'A/F', 'A', 'F'}

\item{deMethod}{DE test method, options are 'wilcox', 't', 'negbinom', 'poisson', 'MAST', 'DESeq2'}

\item{pAdjValCutoff}{adjusted p-value cutoff of significant DEGs}

\item{topNo}{output this number of top up and down significant DEGs in each identified/annotated clusters}

\item{expCondSepName}{character string, user defined name either to be 'org' or any character string}

\item{expCondName2change}{if above 'expCondSepName' is defined not as 'org', provide the name to be changed}
}
\value{
a list item including 3 elements:
1: 'clusterDeMarkers'
2. 'clusterDeResSummary'
3. 'clusterTopDeMarkers'
}
\description{
getClusterExpCondDe() Function
}
\details{
This function is used to identify DEGs in each identified/annotated cell cluster based on orignial/redesigned experimental conditions originated from sample name
}
\examples{
getClusterExpCondDe()
}
\keyword{FindMarkers}
