% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcObjFnPseudo.R
\name{getClusterPseudo}
\alias{getClusterPseudo}
\title{getClusterPseudo() Function}
\usage{
getClusterPseudo(resDir, newAnnotation, newAnnotationRscriptName, clusterName)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{clusterName}{specified a single or several cluster names used for pseudotime analysis}
}
\value{
a SingleCellExperiment where 3 methods functional pseudotime analysis results are saved in colData
if subclusers included, slingshot  used seurat subclusters
}
\description{
getClusterPseudo() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
getClusterPseudo()
}
\keyword{getClusterPseudo}
