% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterSummaryReplot.R
\name{getClusterSummaryReplot}
\alias{getClusterSummaryReplot}
\title{getClusterSummaryReplot() Function}
\usage{
getClusterSummaryReplot(
  resDir = NULL,
  rdsFname = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondSepName = NULL,
  fpClusterOrder = NULL,
  perClusterOrder = NULL,
  expCondNameOrder = NULL
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{expCondCheck}{character string, user defined name either to be 'org' or any character string}

\item{expCondNameOrder}{whether to change the experimental conditions order in summarized feature plots and percentage bar plots}
}
\value{

}
\description{
getClusterSummaryReplot() Function
}
\details{
This function is used to annotate/assign and merge identified clusters,
}
\examples{
clusterSummary()
}
\keyword{clusterSummary}
