% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpCondClusterPseudotime.R
\name{getExpCondClusterPseudotime}
\alias{getExpCondClusterPseudotime}
\title{getExpCondClusterPseudotime() Function}
\usage{
getExpCondClusterPseudotime(
  resDir = NULL,
  rdsFname = NULL,
  newAnnotation = as.logical("F"),
  newAnnotationRscriptName = NULL,
  expCondName = NULL,
  expCondName2change = NULL,
  expCond = NULL,
  cellcluster = NULL
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{clusterName}{specified a single or several cluster names used for pseudotime analysis}
}
\value{
a SingleCellExperiment where 3 methods functional pseudotime analysis results are saved in colData
}
\description{
getExpCondClusterPseudotime() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
getExpCondClusterPseudotime()
}
\keyword{getExpCondClusterPseudotime}
