% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoiDotplot.R
\name{getGoiDotplot}
\alias{getGoiDotplot}
\title{getGoiDotplot() Function}
\usage{
getGoiDotplot(
  resDir,
  newAnnotation,
  newAnnotationRscriptName,
  expCondSepName,
  expCondName2change,
  goiFname,
  dotPlotFnamePrefix,
  dotPlotMinExpCutoff,
  dotPlotWidth,
  dotPlotHeight
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{expCondSepName}{character string, user defined name either to be 'org' or any character string}

\item{expCondName2change}{if above 'expCondSepName' is defined not as 'org', provide the name to be changed}

\item{goiFname}{full path of a file name, where a list of marker/features genes provided}

\item{dotPlotFnamePrefix}{dotPlot file name prefix}

\item{dotPlotMinExpCutoff}{dotPlot miniumn expression threshold}

\item{dotPlotWidth}{dotPlot width}

\item{dotPlotHeight}{dotPlot height}
}
\value{
the dotplots of provided GOI(gene of interest) saved in '' inside the provided 'resDir'
}
\description{
getGoiDotplot() Function
}
\details{
This function is used to make dotplot of marker/features genes in provided 'goiFname'.
}
\examples{
getGoiDotplot()
}
\keyword{GoiDotplot}
