% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHippoRes.R
\name{getHippoRes}
\alias{getHippoRes}
\title{getHippoRes() Function}
\usage{
getHippoRes(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  cellcluster = NULL,
  expCond = NULL,
  noClusters = 3,
  sparseMatrix = F,
  initial.label.on = F,
  hippoResNamePrefix = "hippo_cluster_test",
  topN = 100
)
}
\arguments{
\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{rds}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample or expCond1/2/....}

\item{cellcluster}{specify cell clusters (idents) to be conducted with hippo analysis.}

\item{expCond}{specify the specific experimental conditions to be conducted with hippo analysis.}

\item{noClusters}{number of clusters for hippo to identify, default = 3.}

\item{sparseMatrix}{whether to turn sparseMatrix option on, default off, when turned on, takes longer time to run hippo.}

\item{initial.label.on}{whether to run hippo starting with initial identified cell clusters information.}

\item{hippoResNamePrefix}{prefix of the hippo analysis results, if not defined, by default = 'hippo_cluster_test'.}

\item{topN}{= 100 by default, the top number genes plotted in hippo analysis}
}
\value{
a results directory/folder with saved Rdata hippo analysis results inside
\subsection{------------------------------------------------------------------------------------}{
}
}
\description{
getHippoRes() Function
}
\details{
This function is used to run hippo analysis and save hippo results into Rdata.
}
\examples{
getHippoRes(rds,  expCondCheck='sample/expCond*', cellcluster)
}
\keyword{getHippoRes}
