% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntegrationClusterMarkers.R
\name{getIntegrationClusterMarkers}
\alias{getIntegrationClusterMarkers}
\title{getIntegrationClusterMarkers() Function}
\usage{
getIntegrationClusterMarkers(
  qcProcessedResults,
  resDirName = NULL,
  heatmapTopno = 10
)
}
\arguments{
\item{qcProcessedResults}{returned 'qcProcessObj' or full results from countReadin() function.}

\item{resDirName}{optional, define folder/directory name where integration analysis results will be saved}

\item{heatmapTopno}{TOBEADD}
}
\value{
a list item including 3 elements:
\enumerate{
\item 'integratedObj': finalized integrated seurat object;
\item 'posMarkers': identified positively expressed cluster markers;
\item 'resDir': full path of results directory, where includes TO BE ADDED.
}
}
\description{
getIntegrationClusterMarkers() Function
}
\details{
This function is used to integrate seurat objects with identified anchor via seurat method,
followed with conducing cluster analysis, and identifying cluster markers
}
\examples{
getIntegrationClusterMarkers()
}
\keyword{getIntegrationClusterMarkers,}
\keyword{seuratIntegrate}
