% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScatterPlotInput.R
\name{getScatterPlot}
\alias{getScatterPlot}
\title{getExpCondClusterMarkers() Function}
\usage{
getScatterPlot(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  subsetOn = NULL,
  subset = NULL,
  subsetGenes = NULL,
  expCondCheck = "idents",
  selectedGroups,
  interactive = T
)
}
\arguments{
\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{rds}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{subsetOn}{whether to subset certain cell clusters when it is 'idents' or conditions when it is 'expCond1/2...'.}

\item{subset}{define the specific cell cluster names when it is 'subsetOn=idents' or expConds inheritated from metadata table 'subsetOn=expCond1/2...' to extract.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{selectedGroups}{specify 2 groups in a list to be plotted with scatter plot, it can be either cell ident names presented in 'Idents' of RDS object or experimental condition levels shown in the 'expCondCheck' condition.}
}
\value{
an interactive scatter plot window
}
\description{
getExpCondClusterMarkers() Function
}
\details{
This function is used to identify positively expressed cluster markers for all originally identified/annotated cell clusters from the experimental conditions specified in the metadata table.
}
\examples{
getScatterPlot(rds, expCondCheck='Idents', selectedGroups = list('group1'=c('Ident1', 'Ident2'), 'group2' = c('Ident4', 'Ident5')))
}
\keyword{getScatterPlot}
