% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Improved_Seurtat_Pre_Process.R
\name{Improved_Seurat_Pre_Process}
\alias{Improved_Seurat_Pre_Process}
\title{Improved_Seurat_Pre_Process() Function}
\usage{
Improved_Seurat_Pre_Process(
  seuratObject,
  num_genes = 50,
  write_files = FALSE,
  data_type = "counts"
)
}
\arguments{
\item{seuratObject}{list including full path to cellranger analysis results for different samples.}

\item{num_genes}{num_genes}

\item{write_files}{to be added}

\item{data_type}{to be added}
}
\description{
This function allows you to find and estimate doublets with DoubletDecon for the provide input 'cellrangerResList', taken from DoubletDecon directly due to function calling problems.
}
\examples{
Improved_Seurat_Pre_Process()

}
\keyword{DoubletDecon}
