% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata2listInput.R
\name{meata2list}
\alias{meata2list}
\title{meata2list() Function}
\usage{
meata2list(metadata)
}
\arguments{
\item{metadata}{list including full path to cellranger analysis results for different samples.}
}
\value{

}
\description{
This function allows you to find and estimate doublets with DoubletDecon for the provide iput 'cellrangerResList'
}
\examples{
meata2list()
}
\keyword{meata2list}
