% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdHeatmap.R
\name{pdHeatmap}
\alias{pdHeatmap}
\title{pdHeatmap() Function modified from heatmap.3 to include extra ColSideColors factors.}
\usage{
pdHeatmap(
  x,
  Rowv = TRUE,
  Colv = if (symm) "Rowv" else TRUE,
  distfun = dist,
  hclustfun = hclust,
  dendrogram = c("both", "row", "column", "none"),
  symm = FALSE,
  scale = c("none", "row", "column"),
  na.rm = TRUE,
  revC = identical(Colv, "Rowv"),
  add.expr,
  breaks,
  symbreaks = max(x < 0, na.rm = TRUE) || scale != "none",
  col = "heat.colors",
  colsep,
  rowsep,
  sepcolor = "white",
  sepwidth = c(0.05, 0.05),
  cellnote,
  notecex = 1,
  notecol = "cyan",
  na.color = par("bg"),
  trace = c("none", "column", "row", "both"),
  key.title = NULL,
  key.xlab = NULL,
  tracecol = "cyan",
  hline = median(breaks),
  vline = median(breaks),
  linecol = tracecol,
  margins = c(5, 5),
  ColSideColors,
  RowSideColors,
  side.height.fraction = 0.3,
  cexRow = 0.2 + 1/log10(nr),
  cexCol = 0.2 + 1/log10(nc),
  labRow = NULL,
  labCol = NULL,
  key = TRUE,
  keysize = 1.5,
  density.info = c("none", "histogram", "density"),
  denscol = tracecol,
  symkey = max(x < 0, na.rm = TRUE) || symbreaks,
  densadj = 0.25,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  lmat = NULL,
  lhei = NULL,
  lwid = NULL,
  NumColSideColors = 1,
  NumRowSideColors = 1,
  KeyValueName = "Value",
  ...
)
}
\arguments{
\item{x}{to ADD}

\item{Rowv}{to ADD}

\item{Colv}{determines if and how the \emph{column} dendrogram should be
    reordered.  Has the same options as the \code{Rowv} argument above and
    \emph{additionally} when \code{x} is a square matrix, \code{Colv =
      "Rowv"} means that columns should be treated identically to the
    rows (and so if there is to be no row dendrogram there will not be a
    column one either).}

\item{distfun}{to ADD}

\item{hclustfun}{to ADD}

\item{dendrogram}{to ADD}

\item{symm}{logical indicating if \code{x} should be treated
    \bold{symm}etrically; can only be true when \code{x} is a square matrix.}

\item{scale}{character indicating if the values should be centered and
    scaled in either the row direction or the column direction, or
    none.  The default is \code{"row"} if \code{symm} false, and
    \code{"none"} otherwise.}

\item{na.rm}{logical indicating whether \code{NA}'s should be removed.}

\item{revC}{logical indicating if the column order should be
    \code{\link{rev}}ersed for plotting, such that e.g., for the
    symmetric case, the symmetry axis is as usual.}

\item{add.expr}{expression that will be evaluated after the call to
    \code{image}.  Can be used to add components to the plot.}

\item{margins}{numeric vector of length 2 containing the margins
    (see \code{\link{par}(mar = *)}) for column and row names, respectively.}

\item{ColSideColors}{(optional) character vector of length \code{ncol(x)}
    containing the color names for a horizontal side bar that may be used to
    annotate the columns of \code{x}.}

\item{RowSideColors}{(optional) character vector of length \code{nrow(x)}
    containing the color names for a vertical side bar that may be used to
    annotate the rows of \code{x}.}

\item{cexRow, cexCol}{positive numbers, used as \code{cex.axis} in
    for the row or column axis labeling.  The defaults currently only
    use number of rows or columns, respectively.}

\item{labRow, labCol}{character vectors with row and column labels to
    use; these default to \code{rownames(x)} or \code{colnames(x)},
    respectively.}

\item{main, xlab, ylab}{main, x- and y-axis titles; defaults to none.}

\item{...}{additional arguments passed on to \code{\link{image}},
    e.g., \code{col} specifying the colors.}
}
\description{
pdHeatmap() Function modified from heatmap.3 to include extra ColSideColors factors.
}
\details{
Intermediate function to make heat-map of pseudo-time trajectory analysis results
}
\examples{
pdHeatmap()
}
\keyword{pdHeatmap}
