% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotObjPseudotime.R
\name{plotObjPseudotime}
\alias{plotObjPseudotime}
\title{plotPseudotime() Function}
\usage{
plotObjPseudotime(getClusterPseudoRes)
}
\arguments{
\item{getClusterPseudoRes}{required, results from fn getClusterPseudo()}
}
\value{
2 plots (correlation plots and 3 combined pseudotime plots) with this specified cluster name in input 'getClusterPseudoRes$clusterName'
\subsection{------------------------------------------------------------------------------------}{

Note: PCA, diffusion map used ranked results as pseudotime,
whereas slingshot used slingshot() returned results - column 'slingPseudotime_1'.
row1: 3 methods PC1 vs PC2 pseudotime plot
row2: 3 methods yCol/colCol on rank(PC1) or slingPseudotime_1
correlation plot of 3 methods
}
}
\description{
plotPseudotime() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
plotPseudotime()
}
\keyword{plotPseudotime}
