% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPseudotime.R
\name{plotPseudotime}
\alias{plotPseudotime}
\title{plotPseudotime() Function}
\usage{
plotPseudotime(sceObj2plot, yCol, yLabel, colCol, colLabel, plotFnamePrefix)
}
\arguments{
\item{sceObj2plot}{required, returned results from above fn calcFnPseudo()}

\item{yCol}{required, which column in the provided object needs to be visualized}

\item{yLabel}{optional, y axis lable, by default = 'yCol'}

\item{colCol}{required, which column in the provided object needs to be visualized via different colour}

\item{colLabel}{optional, legend color lable, by default = 'colCol'}

\item{plotFnamePrefix}{optional, returned 2 plots (correlation plots and 3 combined pseudotime plots) with this defined 'plotFnamePrefix'}
}
\value{
2 plots (correlation plots and 3 combined pseudotime plots) with this defined 'plotFnamePrefix'
\subsection{------------------------------------------------------------------------------------}{

Note: PCA, diffusion map used ranked results as pseudotime,
whereas slingshot used slingshot() returned results - column 'slingPseudotime_1'.
}
}
\description{
plotPseudotime() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
plotPseudotime()
}
\keyword{plotPseudotime}
