% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPseudotimeHeatmap.R
\name{plotPseudotimeHeatmap}
\alias{plotPseudotimeHeatmap}
\title{plotPseudotimeHeatmap() Function}
\usage{
plotPseudotimeHeatmap(pseudoRes, plotname = "TEST")
}
\arguments{
\item{pseudoRes}{required, returned results from above fn calcFnPseudo()$pseudoRes$sceObj}

\item{plotname}{if not provided, named as 'test.pdf'}
}
\value{
a plot with corresponding specified x, y-axis used for plot
\subsection{------------------------------------------------------------------------------------}{
}
}
\description{
plotPseudotimeHeatmap() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
plotPseudotimeHeatmap()
}
\keyword{plotPseudotimeHeatmap}
