% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPseudotimeLineages.R
\name{plotPseudotimeLineages}
\alias{plotPseudotimeLineages}
\title{plotPseudotime() Function}
\usage{
plotPseudotimeLineages(pseudoRes, plotname = "TEST")
}
\arguments{
\item{pseudoRes}{required, returned results from above fn calcFnPseudo()$pseudoRes}
}
\value{
noting, but will save a PCA plot with slingshot calculated lineage in defined plotname.pdf
\subsection{------------------------------------------------------------------------------------}{
}
}
\description{
plotPseudotime() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
plotPseudotimeLineages()
}
\keyword{plotPseudotimeLineages}
