% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCellphoneDBinput.R
\name{prepCellphoneDBinput}
\alias{prepCellphoneDBinput}
\title{getExpCondClusterPseudotime() Function}
\usage{
prepCellphoneDBinput(
  rds,
  newAnnotation = "F",
  newAnnotationRscriptName = NULL,
  resDir = NULL,
  resFnamePrefix = "testInput",
  expCondCheck = "sample",
  expCond = NULL,
  cellcluster = NULL,
  debug = as.logical(F)
)
}
\arguments{
\item{rds}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{resDir}{full path of folder directory where results to be saved.}

\item{resFnamePrefix}{result file name prefix.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample or expCond1/2/... presented in the input RDS metadata table column names.}

\item{expCond}{specify the specific experimental condition for pseudo time trajectory analysis, if not specified, all experimental conditions will be performed.}

\item{cellcluster}{specify the specific cell cluster name for pseudo time trajectory analysis, if not specified, all cell clusters will be performed.}

\item{debug}{whether to turn on for debug check, by default FALSE (turned off).}
}
\value{
a SingleCellExperiment where 3 methods functional pseudotime analysis results are saved in colData
}
\description{
getExpCondClusterPseudotime() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
getExpCondClusterPseudotime()
}
\keyword{getExpCondClusterPseudotime}
