% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHippoIdents.R
\name{updateHippoIdents}
\alias{updateHippoIdents}
\title{updateHippoIdents() Function}
\usage{
updateHippoIdents(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  resSaveFname = NULL,
  hippoResList,
  hippoResNames = NULL,
  hippoResK,
  lightHippo = T,
  debug = F
)
}
\arguments{
\item{resDir}{specify a exiting full path of integration results analysis are saved.}

\item{rds}{provide integrated RDS object, user can also provide the full path of the RDS where integrated RDS object is saved.}

\item{newAnnotation}{logical option, whether to add the new cell types annotation for identified cell clusters from provided integrated RDS file.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify the full path of the R script where new cell annotations are defined.}

\item{resSaveFname}{whether to save the seurat object with the updated idents, by default it is 'NULL' without saving idents updated seurat object.}

\item{hippoResList}{list of Rdata inputs for cell types to be updated with HIPPO results}

\item{hippoResNames}{corresponding updated HIPPO results names with respect to list input in 'hippoResList'}

\item{hippoResK}{corresponding cluster k number with respect to list input in 'hippoResList'}

\item{debug}{whether to turn on for debug check, by default FALSE (turned off).}
}
\value{
seurat object with updated idents from hippo results
}
\description{
updateHippoIdents() Function
}
\details{
This function is used to update seurat object's idents based on provided hippo clustering results.
}
\examples{
updateHippoIdents(rds, hippoResList)
}
\keyword{updateHippoIdents}
