% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateIdents.R
\name{updateIdents}
\alias{updateIdents}
\title{updateIdents() Function}
\usage{
updateIdents(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  resSaveFname = NULL,
  hippoResList,
  hippoResNames = NULL,
  hippoResK,
  lightHippo = T
)
}
\arguments{
\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{rds}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{resSaveFname}{whether to save the seurat object with the updated idents, by default it is 'NULL' without saving idents updated seurat object.}
}
\value{
seurat object with updated idents from hippo results
}
\description{
updateIdents() Function
}
\details{
This function is used to update seurat object's idents based on provided hippo clustering results.
}
\examples{
updateIdents()
}
\keyword{updateIdents}
