% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFnPseudo.R
\name{calcPCApseudo}
\alias{calcPCApseudo}
\title{calcPCApseudo() Function}
\usage{
calcPCApseudo(obj, slingshotclusterLabels, resSave, resFnamePrefix)
}
\arguments{
\item{obj}{required, either a seurat or SingleCellExperiment object}

\item{slingshotclusterLabels}{optional, default no cluster used for slingshot}

\item{resSave}{optional, default is F}

\item{resFnamePrefix}{optional, default is 'TEST', when resSave is on}
}
\value{
a SingleCellExperiment where 3 methods functional pseudotime analysis results are saved in colData
}
\description{
calcPCApseudo() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
calcPCApseudo()
}
\keyword{calcPCApseudo}
