% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFnPseudo.R
\name{calcPCApseudo}
\alias{calcPCApseudo}
\title{calcPCApseudo() Function}
\usage{
calcPCApseudo(
  obj,
  slingshotclusterLabels = "GMM",
  topFeatureNo = 2000,
  resSave = "F",
  resFnamePrefix = "TEST"
)
}
\arguments{
\item{obj}{either a Seurat or SingleCellExperiment object.}

\item{slingshotclusterLabels}{cluster used for slingshot with 3 options: NULL, GMM, Seurat_clusters, by default NULL.}

\item{resSave}{whether to save the analysis results into Rdata, by default 'FALSE'}

\item{resFnamePrefix}{if 'resSave = T', the Rdata file name prefix, by default 'TEST'}
}
\value{
a SingleCellExperiment where 3 methods functional pseudo time analysis results are saved in colData
}
\description{
calcPCApseudo() Function
}
\details{
This function is used to perform functional pseudo time analysis via PCA, Diffusion Map, and slingshot
}
\examples{
calcPCApseudo()
}
\keyword{calcPCApseudo}
