% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compAnno.R
\name{compAnno}
\alias{compAnno}
\title{compAnno() Function}
\usage{
compAnno(seuratObj, seuratObj2Comp)
}
\arguments{
\item{seuratObj}{the ident cells to be decomposed for comparisons.}

\item{seuratObj2Comp}{the annotation cells to be compared with.}
}
\value{
results in cell number and relative percentage of idents cells between 2 input seurat objects, all cells are decomposed on rows.
}
\description{
compAnno() Function
}
\details{
This function is used to make comparisons between 2 seurat objects. 'seuratObj2Comp' is the baseline, where 'seuratObj' is decomposed into.
}
\examples{
compAnno()
}
\keyword{compAnno}
