% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDoublets.R
\name{findDoublets}
\alias{findDoublets}
\title{findDoublets() Function}
\usage{
findDoublets(
  metadata,
  multiomics = F,
  extraFilter = F,
  genomeSpecies = NULL,
  doubletDeconRhop = 0.5,
  doubletDeconPMF = F,
  doubletDeconNoCore = -1,
  resFilename = NULL
)
}
\arguments{
\item{metadata}{required, metadata table with at least 2 required columns: 'sample' and 'path', where path shows the count table matrix is located.}

\item{multiomics}{optional, default = F}

\item{extraFilter}{optional, default = F, if 'TRUE', column 'filterFname' needs to be included in the provided metadata table.}

\item{genomeSpecies}{genome species, currently supporting human, mouse, and rate.}

\item{doubletDeconRhop}{option used for DoubletDecon detection, by default = 0.5.}

\item{doubletDeconPMF}{option used for DoubletDecon detection, by default 'FALSE'.}

\item{doubletDeconNoCore}{option used for DoubletDecon detection, by default '-1' no parallel.}

\item{resFilename}{define the folder/directory name where doublets detection results will be saved, if not defined, by default results will be saved at the current working directory in a folder named as 'doublets_results'.}
}
\value{
generate a result directory/folder specified with 'resFilename' to save
doublets identification analysis results for all samples in the metadata with a data structure shown as below:
\enumerate{
\item 'metadata$sample' + '_medoids_doublet_cells_name.Rdata': detected doublet cells with mediods, object named as 'centroidsDoubletCells'
\item 'metadata$sample' + '_centroids_doublet_cells_name.Rdata': detected doublet cells with centroids, object named as 'centroidsDoubletCells'
\item 'metadata$sample' + '_OL_doublet_cells_name.Rdata': detected overlapped doublet cells with centroids and mediods, object named as 'olDoubletCells'
\item 'metadata$sample' + '_doubletDeconRes.Rdata': Main_Doublet_Decon() analysis results for Centroids (object named as 'doubletDeconResCentroids') and Medoids (object named as 'doubletDeconResMedoids')
\item 'metadata$sample' + '_doubletDecon_mediod_centroid_vennComp.pdf': venn-diagram plot of detected doublets with centroids and mediods algorithms
\item 'metadata$sample' + '_doublets_no_summary.txt': centroids and mediods algorithms detected doublets summary
}
}
\description{
This function allows you to identify doublets with DoubletDecon for the provide input in metadata
}
\examples{
findDoublets()
}
\keyword{findDoublets}
