% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterExpCondDE.R
\name{getClusterExpCondDe}
\alias{getClusterExpCondDe}
\title{getClusterExpCondDe() Function}
\usage{
getClusterExpCondDe(
  resDir = NULL,
  rdsFname = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondSepName = NULL,
  expCondName2change = NULL,
  compGroup,
  deMethod = "wilcox",
  pAdjValCutoff = 0.05,
  topNo = 10
)
}
\arguments{
\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{rdsFname}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{expCondCheck}{3 options: 'sample', 'expCond1', or 'expCond2' to specify which experimental conditions to be explored with this function.}

\item{expCondSepName}{part of file name string to specify the analysis results folder name.}

\item{expCondName2change}{character string to indicate part of characters specified here can be removed from sample name defined in the metadata table, if additional samples combination needs to be explored which has not been specified in the column of 'expCond1' or 'expCond2'.}

\item{compGroup}{specify 2 group names for comparisons, these group names is corresponding to experimental condition factor levels.}

\item{deMethod}{DE test method with options: 'wilcox', 't', 'negbinom', 'poisson', 'MAST', 'DESeq2', by default = 'wilcox'.}

\item{pAdjValCutoff}{adjusted p-value cutoff for significant DEGs detection, by default = 0.05.}

\item{topNo}{specify the top number of up and down significant DEGs in each identified/annotated cell clusters, by default = 10.}
}
\value{
a list item including 3 elements:
1: 'clusterDeMarkers'
2. 'clusterDeResSummary'
3. 'clusterTopDeMarkers'
}
\description{
getClusterExpCondDe() Function
}
\details{
This function is used to identify DEGs in each identified/annotated cell cluster based on the experimental conditions from metadata table column 'sample', 'expCond1', or 'expCond2'
}
\examples{
getClusterExpCondDe()
}
\keyword{FindMarkers}
