% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterExpCondDE.R
\name{getClusterExpCondDe}
\alias{getClusterExpCondDe}
\title{getClusterExpCondDe() Function}
\usage{
getClusterExpCondDe(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  cellcluster = NULL,
  compGroup,
  deMethod = "wilcox",
  covariateVarName = NULL,
  deseq2bulk.metaCovariateInput = NULL,
  covariateVarLevels = NULL,
  norm.method = "UQ",
  run.dispersion = as.logical(T),
  min.cells.group = 10,
  min.pct = 0.1,
  logfc.threshold = 0.25,
  pAdjValCutoff = 0.05,
  topNo = 10,
  debug = F,
  outputExcel = as.logical(T)
)
}
\arguments{
\item{resDir}{specify an exiting full path of directory, where results will be saved.}

\item{rds}{provide integrated RDS object, user can also provide the full path of the RDS where integrated RDS object is saved with above rdsDir option.}

\item{newAnnotation}{logical option, whether to add the new cell types annotation for identified cell clusters from provided integrated RDS file.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify the full path of the R script where new cell annotations are defined.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{expCondCheckFname}{suffix of the directory/folder and file name of the dot plot to be saved, if not defined, the same as the 'expCondCheck' option.}

\item{cellcluster}{specify the specific cell cluster names for DE analysis, if not specified, all cell clusters will be performed .}

\item{compGroup}{specify either 1 or 2 group names seperated by '/' for comparisons, these group names is corresponding to experimental condition factor levels. If only 1 group named is specified, it makes comparison between this specified group with all others.}

\item{deMethod}{DE test method with options: 'wilcox', 't', 'negbinom', 'poisson', 'MAST', 'DESeq2', 'DESeq2.bulk', by default = 'wilcox'.}

\item{covariateVarName}{this option is only used when deMethod = 'MIST' with this specified column for covariate correction.}

\item{deseq2bulk.metaCovariateInput}{when deMethod = 'DESeq2.bulk', use this option to provide the meta data table for covariate correction. This input should be a data frame object with all corresponding sample items in rows and corresponding group information on columns.}

\item{norm.method}{DESeq2 counts across samples normalization method, options are TMM, UQ or DEseq2.}

\item{run.dispersion}{whether to run DESeq2 counts across genes dispersion normalization, if norm.method='DEseq2', this option is automatically on.}

\item{min.cells.group}{Minimum number of cells in one of the comparison groups.}

\item{min.pct}{only test genes that are detected in this specified minimum fraction of cells in either of these 2 comparison populations, default is 0.1.}

\item{logfc.threshold}{only test genes that are detected in this specified X-fold difference (log-scale) between these 2 comparison populations cells, default is 0.25 (around 1.19 FC).}

\item{pAdjValCutoff}{adjusted p-value cutoff for significant DEGs detection, by default = 0.05.}

\item{topNo}{specify the top number of up and down significant DEGs in each identified/annotated cell clusters, by default = 10.}

\item{debug}{whether to turn on for debug check, by default FALSE (turned off).}
}
\value{
a list item including 3 elements:
1: 'clusterDeMarkers'
2. 'clusterDeResSummary'
3. 'clusterTopDeMarkers'
}
\description{
getClusterExpCondDe() Function
}
\details{
This function is used to identify DEGs in each identified/annotated cell cluster based on the experimental conditions from metadata table column 'sample', 'expCond1', or 'expCond2'
}
\examples{
getClusterExpCondDe()
}
\keyword{FindMarkers}
