% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterMarkers.R,
%   R/getIntegrationClusterMarkers.R
\name{getClusterMarkers}
\alias{getClusterMarkers}
\title{getClusterMarkers() Function}
\usage{
getClusterMarkers(
  qcProcessedResults,
  integrationMethod = "CCA",
  nfeatures = 2000,
  resDirName = NULL,
  topN = 10
)

getClusterMarkers(
  qcProcessedResults,
  integrationMethod = "CCA",
  nfeatures = 2000,
  resDirName = NULL,
  topN = 10
)
}
\arguments{
\item{qcProcessedResults}{required, processQC() returned full result list or object 'qcProcessObj'.}

\item{integrationMethod}{default = 'CCA', either 'CCA' or 'RPCA' integration method can be applied for anchor identification for integration.}

\item{resDirName}{optional, define folder/directory name where integration analysis results will be saved.
If processQC() full results used for option 'qcProcessedResults', it will use the same 'resDirName' used in processQC()}

\item{topN}{optional, default = 10, indicating the number of identified cluster gene markers}

\item{seuratObjList}{returned 'seuratObjList' from countReadin() function.}

\item{anchorIntegrate}{whehter to implement anchorIntegrate, default = T.}
}
\value{
list item including 3 elements: full path of results directory in 'resDir',
finalized integrated seurat object in 'integratedObj',
and identified positively expressed cluster markers in 'posMarkers'

the entire integration analysis results in the defined 'resDirName' inherited from 'processQC()'.
Additionally, a list item including 3 elements:
\enumerate{
\item 'integratedObj': integrated Seurat object;
\item 'posMarkers': identified positively expressed cluster gene markers;
\item 'resDir': full path of results directory, where the entire integration analysis results are saved.
}
}
\description{
getClusterMarkers() Function

getClusterMarkers() Function
}
\details{
This function is used to integrate seurat objects with identified anchor via seurat method,
followed with conducing cluster analysis, and identifying cluster markers

This function is used to integrate Seurat objects with identified anchor via Seurat method,
followed with clustering analysis, and cluster markers identification.
}
\examples{
getClusterMarkers()
getClusterMarkers()
}
\keyword{getClusterMarkers,}
\keyword{seuratIntegrate}
