% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterMarkers.R
\name{getClusterMarkers}
\alias{getClusterMarkers}
\title{getClusterMarkers() Function}
\usage{
getClusterMarkers(seuratObjList, anchorIntegrate, resDirName)
}
\arguments{
\item{seuratObjList}{returned 'seuratObjList' from countReadin() function.}

\item{anchorIntegrate}{whehter to implement anchorIntegrate, default = T.}

\item{resDirName}{optional, define folder/directory name where integration analysis results will be saved}
}
\value{
list item including 3 elements: full path of results directory in 'resDir',
finalized integrated seurat object in 'integratedObj',
and identified positively expressed cluster markers in 'posMarkers'
}
\description{
getClusterMarkers() Function
}
\details{
This function is used to integrate seurat objects with identified anchor via seurat method,
followed with conducing cluster analysis, and identifying cluster markers
}
\examples{
getClusterMarkers()
}
\keyword{getClusterMarkers,}
\keyword{seuratIntegrate}
