% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntegrationClusterMarkers.R
\name{getClusterMarkers}
\alias{getClusterMarkers}
\title{getClusterMarkers() Function}
\usage{
getClusterMarkers(
  qcProcessedResults,
  integrationMethod = "CCA",
  nfeatures = 2000,
  resDirName = NULL,
  topN = 10,
  ribo = F,
  int.k.weight = 100
)
}
\arguments{
\item{qcProcessedResults}{pre-processed QC results object by processQC().}

\item{integrationMethod}{specify integration method applied for anchor identification for integration, 2 options, 'CCA' and 'RPCA', by default = 'CCA'.}

\item{resDirName}{specify the folder/directory name where integration analysis results will be saved, by default: the save directory where QC results are saved.}

\item{topN}{specify the number of identified over expressed genes from each identified cell clusters, by default = 10.}

\item{ribo}{logical option (True or False), if True, rRNA genes are removed from the variable features for integration analysis.}

\item{int.k.weight}{specify the number of neighbors to consider when weighting anchors in the step of integration, by default = 100, if too few number cells used in the study, this number can be reduced.
If processQC() full results used for option 'qcProcessedResults', it will use the same 'resDirName' used in processQC()}
}
\value{
the entire integration analysis results in the defined 'resDirName' inherited from 'processQC()'.
Additionally, a list item including 3 elements:
\enumerate{
\item 'integratedObj': integrated Seurat object;
\item 'posMarkers': identified positively expressed cluster gene markers;
\item 'resDir': full path of results directory, where the entire integration analysis results are saved.
}
}
\description{
getClusterMarkers() Function
}
\details{
This function is used to integrate Seurat objects with identified anchor via Seurat method,
followed with clustering analysis, and cluster markers identification.
}
\examples{
getClusterMarkers(qcProcessedResults)
}
\keyword{getClusterMarkers,}
\keyword{seuratIntegrate}
