% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterSummaryReplot.R
\name{getClusterSummaryReplot}
\alias{getClusterSummaryReplot}
\title{getClusterSummaryReplot() Function}
\usage{
getClusterSummaryReplot(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  fpRemake = T,
  perRemake = T,
  colors = NULL,
  cellcluster = NULL,
  expCondCheck = "sample",
  expCondSepName = NULL,
  expCondName2change = NULL,
  fpClusterOrder = NULL,
  perClusterOrder = NULL,
  perPlotVertical = F,
  perPlotHeight = NULL,
  perPlotWidth = NULL,
  expCondNameOrder = NULL
)
}
\arguments{
\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{rds}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{fpRemake}{whether to re-make feature plots, default is True}

\item{perRemake}{whether to re-calculate cell number percentage and related percentage plots, default is True}

\item{colors}{by default it is null, colors will be automatically selected, otherwise, a vector string of colors can be specified with this option for colors used in the plots.}

\item{cellcluster}{by default it is null, all cell clusters will be plot, otherwise, cell clusters presented in the plots (UMAP/tSNE and percentage plots) can be specified with this option.}

\item{expCondCheck}{3 options: 'sample', 'expCond1', or 'expCond2' to specify which experimental conditions to be explored with this function.}

\item{expCondSepName}{part of file name string to specify the analysis results folder name.}

\item{expCondName2change}{character string to indicate part of characters specified here can be removed from sample name defined in the metadata table, if additional samples combination needs to be explored which has not been specified in the column of 'expCond1' or 'expCond2'.}

\item{fpClusterOrder}{character string of cell clusters orders to specify the cell clusters orders corresponding to the color scheme, if not defined, sorted numerically, alphabetically, or the orders shown in the annotation R script.}

\item{perClusterOrder}{character string of cell clusters orders to specify the cell clusters orders presented in the percentage bar plot, if not defined, sorted numerically or alphabetically.}

\item{expCondNameOrder}{character string of the corresponding experimental condition factor levels' orders presented in the percentage bar plot, if not defined, sorted numerically or alphabetically.}
}
\value{

}
\description{
getClusterSummaryReplot() Function
}
\details{
This function is used to 1). summarize the number of cells in each originally identified/annotated cells clusters; 2). re-make clustering reduction UMAP/tSNE plots; 3). make the cells percentage plot based on experimental condition levels specified in the metadata table.
}
\examples{
clusterSummary()
}
\keyword{clusterSummary}
