% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterSummaryReplot.R
\name{getClusterSummaryReplot}
\alias{getClusterSummaryReplot}
\title{getClusterSummaryReplot() Function}
\usage{
getClusterSummaryReplot(
  resDir,
  newAnnotation,
  newAnnotationRscriptName,
  expCondSepName,
  expCondName2change,
  clusterLevelReorder = T,
  expCondNameReorder = NULL
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{expCondSepName}{character string, user defined name either to be 'org' or any character string}

\item{expCondName2change}{if above 'expCondSepName' is defined not as 'org', provide the name to be changed}

\item{expCondNameReorder}{whether to change the experimental conditions order in summarized feature plots and percentage bar plots}
}
\value{

}
\description{
getClusterSummaryReplot() Function
}
\details{
This function is used to annotate/assign and merge identified clusters,
}
\examples{
clusterSummary()
}
\keyword{clusterSummary}
