% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterSummaryReplot.R
\name{getClusterSummaryReplot}
\alias{getClusterSummaryReplot}
\title{getClusterSummaryReplot() Function}
\usage{
getClusterSummaryReplot(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  fpRemake = T,
  perRemake = T,
  colors = NULL,
  cellcluster = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  fpClusterOrder = NULL,
  perClusterOrder = NULL,
  perPlotVertical = F,
  perPlotHeight = NULL,
  perPlotWidth = NULL,
  expCondNameOrder = NULL,
  stack = as.logical("T"),
  gap = 0.8,
  debug = as.logical(F)
)
}
\arguments{
\item{resDir}{specify an exiting full path of directory, where results will be saved.}

\item{rds}{provide integrated RDS object, user can also provide the full path of the RDS where integrated RDS object is saved with above rdsDir option.}

\item{newAnnotation}{logical option, whether to add the new cell types annotation for identified cell clusters from provided integrated RDS file.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify the full path of the R script where new cell annotations are defined.}

\item{fpRemake}{whether to re-make feature plots, default is True}

\item{perRemake}{whether to re-calculate cell number percentage and related percentage plots, default is True}

\item{colors}{by default it is null, colors will be automatically selected, otherwise, a vector string of colors can be specified with this option for colors used in the plots.}

\item{cellcluster}{by default it is null, all cell clusters will be plot, otherwise, cell clusters presented in the plots (UMAP/tSNE and percentage plots) can be specified with this option.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{expCondCheckFname}{suffix of the directory/folder and file name of the dot plot to be saved, if not defined, the same as the 'expCondCheck' option.}

\item{fpClusterOrder}{character string of cell clusters orders to specify the cell clusters orders corresponding to the color scheme, if not defined, sorted numerically, alphabetically, or the orders shown in the annotation R script.}

\item{perClusterOrder}{character string of cell clusters orders to specify the cell clusters orders presented in the percentage bar plot, if not defined, sorted numerically or alphabetically.}

\item{expCondNameOrder}{character string of the corresponding experimental condition factor levels' orders presented in the percentage bar plot, if not defined, sorted numerically or alphabetically.}

\item{stack}{whether to make stacked bar-plot}

\item{gap}{space of un-stacked bar-plot between different conditions, 0.1-1.}

\item{debug}{logical option, turn on to have extra analysis information printing out for debug, by defalut 'off'.}
}
\description{
getClusterSummaryReplot() Function
}
\details{
This function is used to 1). summarize the number of cells in each originally identified/annotated cells clusters; 2). re-make clustering reduction UMAP/tSNE plots; 3). make the cells percentage plot based on experimental condition levels specified in the metadata table.
}
\examples{
clusterSummary(rds, expCondCheck='sample/idents/expCond*')
}
\keyword{clusterSummary}
