% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpCondClusterMarkers.R
\name{getExpCondClusterMarkers}
\alias{getExpCondClusterMarkers}
\title{getExpCondClusterMarkers() Function}
\usage{
getExpCondClusterMarkers(
  resDir = NULL,
  rdsFname = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondSepName = NULL,
  pAdjValCutoff = 0.05,
  topNo = 10
)
}
\arguments{
\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{rdsFname}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{pAdjValCutoff}{adjusted p-value cutoff for significant positively expressed cluster markers, by default = 0.05.}

\item{topNo}{specify the top number of up and down significant positively expressed cluster markers in each identified/annotated clusters, by default = 10.}
}
\value{
a list item including 3 elements:
}
\description{
getExpCondClusterMarkers() Function
}
\details{
This function is used to identify positively expressed cluster markers for all originally identified/annotated cell clusters from the experimental conditions specified in the metadata table.
}
\examples{
getExpCondClusterMarkers()
}
\keyword{getExpCondClusterMarkers}
