% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterExpCondMarkers.R
\name{getExpCondClusterMarkers}
\alias{getExpCondClusterMarkers}
\title{getExpCondClusterMarkers() Function}
\usage{
getExpCondClusterMarkers(
  resDir,
  newAnnotation,
  newAnnotationRscriptName,
  clusterReOrder,
  reorderedClusters,
  expCondSepName,
  expCondName2change,
  pAdjValCutoff,
  topNo
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{expCondSepName}{character string, user defined name either to be 'org' or any character string}

\item{expCondName2change}{if above 'expCondSepName' is defined not as 'org', provide the name to be changed}

\item{pAdjValCutoff}{adjusted p-value cutoff of significant positively expressed cluster markers}

\item{topNo}{output this number of top up and down significant positively expressed cluster markers in each identified/annotated clusters}
}
\value{
a list item including 3 elements:
}
\description{
getExpCondClusterMarkers() Function
}
\details{
This function is used to identify positively expressed cluster markers in specified expreimental conditions
}
\examples{
getExpCondClusterMarkers()
}
\keyword{getExpCondClusterMarkers}
