% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterPseudotime.R,
%   R/getExpCondClusterPseudotime.R
\name{getExpCondClusterPseudotime}
\alias{getExpCondClusterPseudotime}
\title{getExpCondClusterPseudotime() Function}
\usage{
getExpCondClusterPseudotime(
  resDir = NULL,
  rds = NULL,
  newAnnotation = "F",
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  expCond = NULL,
  cellcluster = NULL,
  slingshotclusterLabels = NULL,
  topFeatureNo = 2000
)

getExpCondClusterPseudotime(
  resDir = NULL,
  rds = NULL,
  newAnnotation = "F",
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  expCond = NULL,
  cellcluster = NULL,
  slingshotclusterLabels = NULL,
  topFeatureNo = 2000
)
}
\arguments{
\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{expCondCheckFname}{suffix of the directory/folder and file name of the dot plot to be saved, if not defined, the same as the 'expCondCheck' option.}

\item{expCond}{specify the specific experimental condition for pseudo time trajectory analysis, if not specified, all experimental conditions will be performed .}

\item{cellcluster}{specify the specific cell cluster name for pseudo time trajectory analysis, if not specified, all cell clusters will be performed .}

\item{rdsFname}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}
}
\value{
a SingleCellExperiment where 3 methods functional pseudotime analysis results are saved in colData

a SingleCellExperiment where 3 methods functional pseudotime analysis results are saved in colData
}
\description{
getExpCondClusterPseudotime() Function

getExpCondClusterPseudotime() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot

This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
getExpCondClusterPseudotime()
getExpCondClusterPseudotime()
}
\keyword{getExpCondClusterPseudotime}
