% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpCondClusterPseudotime.R
\name{getExpCondClusterPseudotime}
\alias{getExpCondClusterPseudotime}
\title{getExpCondClusterPseudotime() Function}
\usage{
getExpCondClusterPseudotime(
  resDir = NULL,
  rds = NULL,
  newAnnotation = "F",
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  expCond = NULL,
  cellcluster = NULL,
  slingshotclusterLabels = "GMM",
  topFeatureNo = 2000
)
}
\arguments{
\item{resDir}{specify an exiting full path of directory, where results will be saved.}

\item{rds}{provide integrated RDS object, user can also provide the full path of the RDS where integrated RDS object is saved with above rdsDir option.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{expCondCheckFname}{suffix of the directory/folder and file name of the dot plot to be saved, if not defined, the same as the 'expCondCheck' option.}

\item{expCond}{specify a specific experimental condition for pseudo time trajectory analysis, if not specified, all experimental conditions will be performed .}

\item{cellcluster}{specify the specific cell cluster name for pseudo time trajectory analysis, if not specified, all cell clusters will be performed.}

\item{slingshotclusterLabels}{specify slighshot cluster label options, 3 options are available here, they are 'NULL', 'seurat_clusters', or 'GMM'. By default is 'GMM'.}

\item{topFeatureNo}{specify the number of features used in PCA pseudo-time trajectory calculation.}
}
\value{
a SingleCellExperiment where 3 methods functional pseudotime analysis results are saved in colData
}
\description{
getExpCondClusterPseudotime() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
getExpCondClusterPseudotime()
}
\keyword{getExpCondClusterPseudotime}
