% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoiDotplot.R
\name{getGoiDotplot}
\alias{getGoiDotplot}
\title{getGoiDotplot() Function}
\usage{
getGoiDotplot(
  resDir = NULL,
  rdsFname = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  goiFname,
  geneCellTypeOrder = NULL,
  expCondCheck = "sample",
  expCondSepName = NULL,
  expCondName2change = NULL,
  expCondReorderLevels = NULL,
  dotPlotFnamePrefix = "goiDotplots",
  dotPlotMinExpCutoff = 0.3,
  dotPlotWidth = NULL,
  dotPlotHeight = NULL
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{goiFname}{full path of a file name, where a list of marker/features genes provided}

\item{expCondSepName}{character string, user defined name either to be 'org' or any character string}

\item{expCondName2change}{if above 'expCondSepName' is defined not as 'org', provide the name to be changed}

\item{expCondReorderLevels}{orders presented on y-axis from bottom to top}

\item{dotPlotFnamePrefix}{dot Plot file name prefix}

\item{dotPlotMinExpCutoff}{dot Plot minimum expression threshold}

\item{dotPlotWidth}{dot Plot width}

\item{dotPlotHeight}{dot Plot height}
}
\value{
the dotplots of provided GOI(gene of interest) saved in '' inside the provided 'resDir'
}
\description{
getGoiDotplot() Function
}
\details{
This function is used to make dotplot of marker/features genes in provided 'goiFname'.
}
\examples{
getGoiDotplot()
}
\keyword{GoiDotplot}
