% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoiDotplot.R
\name{getGoiDotplot}
\alias{getGoiDotplot}
\title{getGoiDotplot() Function}
\usage{
getGoiDotplot(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  goiFname,
  geneTypeOrder = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  expCondReorderLevels = NULL,
  expCond = NULL,
  cellcluster = NULL,
  dotPlotFnamePrefix = "goiDotplots",
  dotPlotMinExpCutoff = 0.3,
  dotPlotMaxExpCutoff = 2.5,
  dotPlotWidth = NULL,
  scale.min = NA,
  scale.max = NA,
  dotPlotHeight = NULL,
  legendPer = NULL,
  genetypebarPer = NULL,
  fontsize.x = 24,
  fontsize.y = 18,
  fontangle.x = 90,
  fontangle.y = 0,
  fontsize.legend1 = 20,
  fontsize.legend2 = NULL,
  gridOn = as.logical(T),
  geneTypeLegendOn = as.logical(T),
  debug = F
)
}
\arguments{
\item{resDir}{specify an exiting full path of directory, where results will be saved.}

\item{rds}{provide integrated RDS object, user can also provide the full path of the RDS where integrated RDS object is saved with above rdsDir option.}

\item{newAnnotation}{logical option, whether to add the new cell types annotation for identified cell clusters from provided integrated RDS file.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify the full path of the R script where new cell annotations are defined.}

\item{goiFname}{path to file, where a list of marker/features genes are provided in column 'Gene', if column 'Cell Type' is also provided, option 'geneTypeOrder' can be used to adjust orders.}

\item{geneTypeOrder}{if column 'Cell Type' is provided in the 'goiFname' file, this option can be used to adjust the orders of marker gene's cell types, if not provided, marker gene's cell types will be sorted alphabetically.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{expCondCheckFname}{suffix of the directory/folder and file name of the dot plot to be saved, if not defined, the same as the 'expCondCheck' option.}

\item{expCondReorderLevels}{a character string of the corresponding experimental condition factor levels' orders presented on the y-axis of the dot-plot from bottom to top, if not defined, sorted numerically or alphabetically.}

\item{expCond}{subset the specified experimental condition corresponding to 'expCondCheck' levels to make the corresponding dot plot.}

\item{cellcluster}{specify cell clusters to be displayed on the dot plot.}

\item{dotPlotFnamePrefix}{prefix of the dot plot file name, if not defined, by default = 'goiDotplots'.}

\item{dotPlotMinExpCutoff}{minimum expression value threshold presented in the dot plot, if not defined, by default = 0.3.}

\item{dotPlotMaxExpCutoff}{maximum expression value threshold presented in the dot plot, if not defined, by default = 2.5.}

\item{dotPlotWidth}{dot plot width, if not defined, will be decided automatically based on the number of marker genes presented in 'goiFname'}

\item{dotPlotHeight}{dot plot height, if not defined, will be decided automatically based on the number of experimental condition or sample's cell clusters.}

\item{legendPer}{specify the legend proportion in the dot plot, if not specified, by default = 0.1}

\item{genetypebarPer}{specify the proportion top gene types color bar, by default 0.01.}

\item{fontsize.x}{specify the font size of gene names on x-axis, by default 24.}

\item{fontsize.y}{specify the font size of samples attributes on y-axis, by default 18.}

\item{fontangle.x}{change the x-axis text label position, by default vertically aligned with x-axis with 90 degree.}

\item{fontangle.y}{change the y-axis text label position, by default horizontally displayed with y-axis with 90 degree.}

\item{fontsize.legend1}{specify the font size of expression color legend bar.}

\item{fontsize.legend2}{specify the font size of input gene list's types legend bar.}

\item{gridOn}{logical option, whether to have grid on the plot background, by default it is on.}

\item{geneTypeLegendOn}{logical option, whether to include feature genes type legend bar in the plot, by default 'on'.}

\item{debug}{logical option, turn on to have extra analysis information printing out for debug, by defalut 'off'.}
}
\value{
the dotplots of provided GOI(gene of interest) saved in '' inside the provided 'resDir'
}
\description{
getGoiDotplot() Function
}
\details{
This function is used to make feature genes dotplot of a set of marker/feature genes provided with 'goiFname' in an excel file.
}
\examples{
getGoiDotplot(rds, goiFname, expCondCheck='sample/idents/expCond*')
}
\keyword{GoiDotplot}
