% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoiFeatureplot.R
\name{getGoiFeatureplot}
\alias{getGoiFeatureplot}
\title{getGoiFeatureplot() Function}
\usage{
getGoiFeatureplot(
  resDir = NULL,
  rdsFname = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondName2change = NULL,
  goiFname,
  featurePlotMinExpCutoff = 0.3,
  featurePlotReductionMethod = "umap",
  featurePlotFnamePrefix = "goiFeaturePlot"
)
}
\arguments{
\item{resDir}{full path of integration results analysis returned in getClusterMarkers()}

\item{newAnnotation}{logical value, whether to provide manual annotation}

\item{newAnnotationRscriptName}{if newAnnotation == T, this script is used to redefine the old clusters}

\item{expCondName2change}{if above 'expCondSepName' is defined not as 'org', provide the name to be changed}

\item{goiFname}{full path of a file name, where a list of marker/features genes provided}

\item{featurePlotMinExpCutoff}{featurePlot miniumn expression threshold}

\item{featurePlotReductionMethod}{featurePlot reduction methods: options are 'tsne' or 'umap'}

\item{featurePlotFnamePrefix}{featurePlot file name prefix}

\item{expCondSepName}{character string, user defined name either to be 'org' or any character string}
}
\value{
the dotplots of provided GOI(gene of interest) saved in '' inside the provided 'resDir'
}
\description{
getGoiFeatureplot() Function
}
\details{
This function is used to make dotplot of marker/features genes in provided 'goiFname'.
}
\examples{
getGoiFeatureplot()
}
\keyword{GoiFeatureplot}
