% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoiFeatureplot.R
\name{getGoiFeatureplot}
\alias{getGoiFeatureplot}
\title{getGoiFeatureplot() Function}
\usage{
getGoiFeatureplot(
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  expCondReorderLevels = NULL,
  cellcluster = NULL,
  goiFname,
  featurePlotMinExpCutoff = 0.3,
  featurePlotMaxExpCutoff = NULL,
  featurePlotReductionMethod = "umap",
  featurePlotLabel = T,
  featurePlotFnamePrefix = "goiFeaturePlot"
)
}
\arguments{
\item{resDir}{specify an exiting full path of directory, where results will be saved.}

\item{rds}{provide integrated RDS object, user can also provide the full path of the RDS where integrated RDS object is saved with above rdsDir option.}

\item{newAnnotation}{logical option, whether to add the new cell types annotation for identified cell clusters from provided integrated RDS file.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify the full path of the R script where new cell annotations are defined.}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{expCondCheckFname}{suffix of the directory/folder and file name of the dot plot to be saved, if not defined, the same as the 'expCondCheck' option.}

\item{cellcluster}{optional, if needed, this option can be used to specify cell clusters to be displayed on the feature plot}

\item{goiFname}{path to file, where a list of marker/features genes are provided in column 'Gene', if column 'Cell Type' is also provided, option 'geneCellTypeOrder' can be used to adjust orders.}

\item{featurePlotMinExpCutoff}{feature plot minimum expression value threshold, by default = 0.3.}

\item{featurePlotReductionMethod}{feature plot clustering reduction methods 'tsne' or 'umap', by default 'umap'.}

\item{featurePlotLabel}{whether to label feature plot cell types, by default True.}

\item{featurePlotFnamePrefix}{feature plot file name prefix, if not defined, by default 'goiFeaturePlot'.}
}
\value{
the dotplots of provided GOI(gene of interest) saved in '' inside the provided 'resDir'
}
\description{
getGoiFeatureplot() Function
}
\details{
This function is used to make dotplot of marker/features genes in provided 'goiFname'.
}
\examples{
getGoiFeatureplot(rds, goiFname, expCondCheck='sample/idents/expCond*')
}
\keyword{GoiFeatureplot}
