% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoiHeatmap.R
\name{getGoiHeatmap}
\alias{getGoiHeatmap}
\title{getGoiHeatmap() Function}
\usage{
getGoiHeatmap(
  heatmap.view = "ident",
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  resSaveFname = NULL,
  scaled = F,
  goiFname = NULL,
  geneNames = NULL,
  expCondCheck = "sample",
  expCondSepName = NULL,
  expCondName2change = NULL,
  cellcluster = NULL,
  expCond = NULL,
  expCondReorderLevels = NULL,
  plotFnamePrefix = "goiHeatmap",
  plotWidth = 25,
  plotHeight = 20,
  minVal = -2.5,
  maxVal = 2.5,
  fontsize.legend = 20,
  fontangle.top = 0,
  fontsize.top = 5,
  fontsize.y = 20,
  fontangle.y = 0,
  hjust.top = 0,
  vjust.top = 0,
  barHeight.top = 0.02
)
}
\arguments{
\item{heatmap.view}{3 options: 'ident', 'expCond', and 'expCond.ident' to make heatmap, by default 'idents'.}

\item{resDir}{full path of integration results analysis are saved, where RDS file is saved inside the 'RDS_Dir'. This path is also returned by getClusterMarkers() execution.}

\item{rds}{User also can provide the full path of RDS file instead of 'resDir' where RDS file is saved in. If this option is used, please also provide 'resDir' to specify where the analysis results will be saved.}

\item{newAnnotation}{logical value to indicate whether to add the annotation for identified cell clusters from getClusterMarkers() integration analysis.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify here for the full path of the R script where cell clusters are defined.}

\item{resSaveFname}{whether to save re-scaled seurat object on the defined 'cellcluster' and 'expCond', by default it is 'NULL' without saving re-scaled RDS object.}

\item{scaled}{whether input rds is scaled, by default False, will scale data based on selected 'cellcluster' and 'expCond'.}

\item{goiFname}{path to file, where a list of marker/features genes are provided in column 'Gene', if column 'Cell Type' is also provided, option 'geneTypeOrder' can be used to adjust orders.}

\item{geneNames}{provide gene names directly with this option, if do not want to use option 'goiFname'}

\item{expCondCheck}{3 options: 'sample', 'expCond1', or 'expCond2' to specify which experimental conditions to be explored with this function.}

\item{expCondSepName}{suffix of the directory/folder and file name of the dot plot to be saved, if not defined, the same as the 'expCondCheck' option.}

\item{expCondName2change}{a character string to indicate part of characters specified here can be removed from sample name defined in the metadata table, if additional samples combination needs to be explored which has not been specified in the column of 'expCond1' or 'expCond2'.}

\item{cellcluster}{specify cell clusters to be displayed on the dot plot}

\item{expCond}{specify the specific experimental conditions to be plotted.}

\item{expCondReorderLevels}{a character string of the corresponding experimental condition factor levels' orders shown on top color bar.}

\item{plotFnamePrefix}{prefix of the heatmap plot file name, if not defined, by default = 'goiHeatmap'.}

\item{plotWidth}{dot plot width, if not defined, will be decided automatically based on the number of marker genes presented in 'goiFname'}

\item{plotHeight}{dot plot height, if not defined, will be decided automatically based on the number of experimental condition or sample's cell clusters.}

\item{fontangle.top}{modify the angle of the text on the top legend color bar.}

\item{fontsize.top}{modify the font size of top legend color bar.}

\item{fontsize.y}{modify the font size of y-axis (gene names).}

\item{fontangle.y}{modify the y-axis (gene names) angle position.}
}
\value{
the Heatmap of provided GOI(gene of interest) provided by 'geneNames' directly or presented in 'goiFname'.
}
\description{
getGoiHeatmap() Function
}
\details{
This function is used to make heat-map of marker/features genes in provided 'goiFname' or geneNames.
}
\examples{
getGoiHeatmap()
}
\keyword{Heatmap}
