% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGoiHeatmap.R
\name{getGoiHeatmap}
\alias{getGoiHeatmap}
\title{getGoiHeatmap() Function}
\usage{
getGoiHeatmap(
  heatmap.view = "idents",
  resDir = NULL,
  rds = NULL,
  newAnnotation = F,
  newAnnotationRscriptName = NULL,
  slot = "scale.data",
  goiFname = NULL,
  geneNames = NULL,
  expCondCheck = "sample",
  expCondCheckFname = NULL,
  cellcluster = NULL,
  expCond = NULL,
  expCondReorderLevels = NULL,
  plotFnamePrefix = "goiHeatmap",
  plotWidth = 25,
  plotHeight = 20,
  res.output = as.logical(F),
  outputFname = NULL,
  minVal = -2.5,
  maxVal = NULL,
  fontsize.legend = 20,
  features.process = as.logical("F"),
  features.process.lists = NULL,
  features.process.fns = "+",
  fontangle.top = 0,
  fontsize.top = 5,
  fontsize.y = 20,
  fontangle.y = 0,
  hjust.top = 0,
  vjust.top = 0,
  barHeight.top = 0.02,
  display = "cell",
  draw.lines = as.logical(T),
  debug = as.logical(F)
)
}
\arguments{
\item{heatmap.view}{specify heatmap visualization options, including 3 options: 'idents', 'expCond', and 'expCond.idents' to make heatmap, by default 'idents'.}

\item{resDir}{specify an exiting full path of directory, where results will be saved.}

\item{rds}{provide integrated RDS object, user can also provide the full path of the RDS where integrated RDS object is saved with above rdsDir option.}

\item{newAnnotation}{logical option, whether to add the new cell types annotation for identified cell clusters from provided integrated RDS file.}

\item{newAnnotationRscriptName}{if 'newAnnotation = T', please specify the full path of the R script where new cell annotations are defined.}

\item{goiFname}{path to file, where a list of marker/features genes are provided in column 'Gene', if column 'Cell Type' is also provided, option 'geneTypeOrder' can be used to adjust orders.}

\item{geneNames}{provide gene names directly with this option, if do not want to use option 'goiFname'}

\item{expCondCheck}{specify which experimental conditions to be explored, including sample, idents, or expCond1/2/....}

\item{expCondCheckFname}{suffix of the directory/folder and file name of the dot plot to be saved, if not defined, the same as the 'expCondCheck' option.}

\item{cellcluster}{specify cell clusters to be displayed on the dot plot}

\item{expCond}{specify the specific experimental conditions to be plotted.}

\item{expCondReorderLevels}{a character string of the corresponding experimental condition factor levels' orders shown on top color bar.}

\item{plotFnamePrefix}{prefix of the heatmap plot file name, if not defined, by default = 'goiHeatmap'.}

\item{plotWidth}{dot plot width, if not defined, will be decided automatically based on the number of marker genes presented in 'goiFname'}

\item{plotHeight}{dot plot height, if not defined, will be decided automatically based on the number of experimental condition or sample's cell clusters.}

\item{fontangle.top}{modify the angle of the text on the top legend color bar.}

\item{fontsize.top}{modify the font size of top legend color bar.}

\item{fontsize.y}{modify the font size of y-axis (gene names).}

\item{fontangle.y}{modify the y-axis (gene names) angle position.}

\item{display}{3 options: 'cell', 'sample.merge' or 'expCond.merge'.}

\item{draw.lines}{whether to have an empty cell space on heatmap to separate different groups.}

\item{debug}{if on, extra debug printing meassages displayed, by default it is off.}
}
\value{
the Heatmap of provided GOI(gene of interest) provided by 'geneNames' directly or presented in 'goiFname'.
}
\description{
getGoiHeatmap() Function
}
\details{
This function is used to make heatmap of marker/features genes in provided 'goiFname' or geneNames with 3 different visualization options on 1). identified cell clusters ('idents'); 2). specific samples attributes from an experimental condition levels ('expCond'); 3). combined levels of cell clusters and samples attributes ('expCond.idents).
}
\examples{
getGoiHeatmap(rds, geneNames/goiFname, expCondCheck='sample/idents/expCond*')
}
\keyword{Heatmap}
