% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIdentPtPlots.R
\name{makeIdentPlots}
\alias{makeIdentPlots}
\title{makeIdentPlots() Function}
\usage{
makeIdentPlots(pseudoRes, plotname = "TEST")
}
\arguments{
\item{pseudoRes}{required, results from fn getClusterPseudo()}

\item{plotname}{pseudo time GMM clustering plot name prefix, by default 'TEST'}
}
\value{
2 plots (correlation plots and 3 combined pseudo time plots) with this specified cluster name in input 'pseudoRes$clusterName'
\subsection{------------------------------------------------------------------------------------}{
}
}
\description{
makeIdentPlots() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
makeIdentPlots()
}
\keyword{makeIdentPlots}
