% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPseudotime.R
\name{plotPseudotime}
\alias{plotPseudotime}
\title{plotPseudotime() Function}
\usage{
plotPseudotime(pseudoRes, xCol, xLabel, yCol, yLabel, colCol, colLabel)
}
\arguments{
\item{pseudoRes}{required, returned results from above fn calcFnPseudo()$pseudoRes}

\item{xCol}{required, which column in the provided object needs to be visualized as x-axis}

\item{xLabel}{optional, x axis lable, by default = 'xCol'}

\item{yCol}{required, which column in the provided object needs to be visualized as y-axis}

\item{yLabel}{optional, y axis lable, by default = 'yCol'}

\item{colCol}{required, which column in the provided object needs to be visualized via different colour}

\item{colLabel}{optional, legend color lable, by default = 'colCol'}
}
\value{
a plot with corresponding specified x, y-axis used for plot
\subsection{------------------------------------------------------------------------------------}{
}
}
\description{
plotPseudotime() Function
}
\details{
This function is used to perform functional pseudotime analysis via PCA, Diffusion Map, and slingshot
}
\examples{
plotPseudotime()
}
\keyword{plotPseudotime}
