% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPseudotimeHeatmap.R
\name{plotPseudotimeHeatmap}
\alias{plotPseudotimeHeatmap}
\title{plotPseudotimeHeatmap() Function}
\usage{
plotPseudotimeHeatmap(pseudoRes, plotname = "TEST")
}
\arguments{
\item{pseudoRes}{required, returned results from function calcFnPseudo()$pseudoRes$sceObj}

\item{plotname}{GAM clustering heatmap plot file name prefix, by default 'TEST_ptGMM_GAMheatmap.pdf'}
}
\value{
a plot with corresponding specified x, y-axis used for plot
\subsection{------------------------------------------------------------------------------------}{
}
}
\description{
plotPseudotimeHeatmap() Function
}
\details{
This function is used to perform functional pseudo time analysis via PCA, Diffusion Map, and slingshot
}
\examples{
plotPseudotimeHeatmap()
}
\keyword{plotPseudotimeHeatmap}
