% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPseudotimeLineages.R
\name{plotPseudotimeLineages}
\alias{plotPseudotimeLineages}
\title{plotPseudotime() Function}
\usage{
plotPseudotimeLineages(pseudoRes, plotname = "TEST")
}
\arguments{
\item{pseudoRes}{required, returned results from function calcFnPseudo()$pseudoRes}

\item{plotname}{GMM slingshot lineage plot file name prefix, by default 'TEST_ptGMM_lineage.pdf'}
}
\value{
generate a PCA plot with slingshot calculated lineage in defined plotname.pdf
\subsection{------------------------------------------------------------------------------------}{
}
}
\description{
plotPseudotime() Function
}
\details{
This function is used to perform functional pseudo time analysis via PCA, Diffusion Map, and slingshot
}
\examples{
plotPseudotimeLineages()
}
\keyword{plotPseudotimeLineages}
