% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReadin.R
\name{processQC}
\alias{processQC}
\title{processQC() Function}
\usage{
processQC(
  metadata,
  resDirName = NULL,
  genomeSpecies = NULL,
  minCells = 3,
  minFeatures = 200,
  mtFiltering = F,
  mtPerCutoff = NULL
)
}
\arguments{
\item{metadata}{required, metadata table with at least 4 required columns: 'sample', 'path', 'doubletsRmMethod', and 'expCond1'.}

\item{resDirName}{define the folder/directory name where integration analysis results will be saved, if not defined, by default results will be saved at the current working directory in a folder named as 'scRICA_results'.}

\item{genomeSpecies}{specify sample's genome species, by default 'human', currently supporting human, mouse, and rate.}

\item{minCells}{the minimum detected number of cells to be included in the analysis used by Seurat, bu default = 3.}

\item{minFeatures}{the minimum number of expressed gene features for cells included in the analysis used by Seurat, by default = 200.}

\item{mtFiltering}{logical to indicate whether to filter mitochondrial content, by default 'FALSE'.}

\item{mtPerCutoff}{if 'mtFiltering = T', this option is required to indicate the percentage cut-off for mitochondrial content filtering.}
}
\value{
the QC analysis results saved in defined 'resDir' under current working directory.
Meanwhile, this will also return a list item including 3 elements:
\enumerate{
\item 'countReadInOjb': a list of Seurat objects of original counts matrix without QC filtering.
\item 'qcProcessObj': a list of Seurat objects with QC passed, such as doublets removal and mitochondrial content removal.
\item 'resDir': full path of the directory/folder name where the QC results are saved.
}
}
\description{
processQC() Function
}
\details{
This function is used to read-in cellranger results into Seurat object by defining w/wo doublets removal.
}
\examples{
processQC()
}
\keyword{processQC}
