% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReadin.R
\name{processQC}
\alias{processQC}
\title{processQC() Function}
\usage{
processQC(
  metadata,
  resDirName = NULL,
  genomeSpecies = NULL,
  minCells = 3,
  minFeatures = 200,
  mtFiltering = F,
  mtPerCutoff = NULL
)
}
\arguments{
\item{metadata}{required, metadata table}

\item{resDirName}{optional, define folder/directory name where integration analysis results will be saved, if not defined, results will be saved current directory in a folder named as 'scRICA_results'}

\item{genomeSpecies}{optional, default = 'human', specify sample's genome species}

\item{minCells}{optional, default = 3, used in creating Seurat object to indicate the minimum detected number of cells to be included.}

\item{minFeatures}{optional, default = 200, used in creating Seurat object to indicate the minimum detected number of gene features (genes) of included cells.}

\item{mtFiltering}{default = F, logical, whether to filter mitochondrial content}

\item{mtPerCutoff}{optional, if mtFiltering = T, default = 20, indicating the percentage cut-off of mitochondrial content}
}
\value{
the QC analysis results saved in defined 'resDir' under current working directory.
Meanwhile, this will also return a list item including 3 elements:
\enumerate{
\item 'countReadInOjb': a list of Seurat objects of original counts matrix without QC filtering.
\item 'qcProcessObj': a list of Seurat objects with QC passed, such as doublets removal and mitochondrial content removal.
\item 'resDir': full path of the directory/folder name where the QC results are saved.
}
}
\description{
processQC() Function
}
\details{
This function is used to readin cellranger results into seurat object by defining w/wo doublets removal
}
\examples{
processQC()
}
\keyword{processQC}
