% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countReadin.R
\name{processQC}
\alias{processQC}
\title{processQC() Function}
\usage{
processQC(
  metadata,
  resDirName = NULL,
  genomeSpecies = NULL,
  minCells = 3,
  minFeatures = 200,
  mtFiltering = F,
  mtPerCutoff = NULL
)
}
\arguments{
\item{metadata}{txt file with 4 columns,
where sample specifiy the sample names,
path specify full path to cellranger analysis results for that sample,
doubletsRmMethods specify doublet removal methods,
doubletsResDir specify the full path of saved doublet removal results.}

\item{resDirName}{optional, define folder/directory name where integration analysis results will be saved}

\item{genomeSpecies}{To be added}

\item{minCells}{optional, default = 3, used in creating seurat object to indicate the minumn detected number of cells to be included.}

\item{minFeatures}{optional, default = 200, used in creating seurat object to indicate the minimum detected number of gene features (genes) of included cells.}

\item{mtFiltering}{default = F, logical, whehter to filter mitochondrial content}

\item{mtPerCutoff}{optional, if mtFiltering = T, default = 20, indicating the percentage cut-off of mitochondrial content}
}
\value{
a list item including 3 elements:
\enumerate{
\item 'countReadInOjb': a list of orignial counts seurat objects
\item 'qcProcessObj': a list of QC passed seurat objects, such as mitochondrial content removal
\item 'resDir': full path of directory/folder name where the entire QC analysis results are saved, it includes
TO BE ADDED.
}
}
\description{
processQC() Function
}
\details{
This function is used to readin cellranger results into seurat object by defining w/wo doublets removal
}
\examples{
processQC()
}
\keyword{processQC}
